Link to statsviewer player (#241)
* parse uri on frontend

* click to copy sv link

* go to statsviewer

* adjust list int. button, add some error handling for banned/nonexistent players

* Fix statsviewer hover styling

Co-authored-by: Patrick <43299462+stadust@users.noreply.github.com>

* remove banned users stuff, add link to statsviewer from records list

* add record video to percentage column

* append player parameter to URL bar on player select

* fix: display error message if ?player=0 on statsviewer

zero is false-y, so it was triggering the path for "malformed player
id". Does not really make a difference because there will never be a
player with id 0, but oh well.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* js: reset pagination error message on selection and navigation

Otherwise it'll just stick around forever, which is a bit annoying if
someone clicks a stats viewer link for a non-existing player.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* minor cleanups

- remove copy-to-clipboard when clicking player name. The on-hover
  underlining was causing the flags to resize, and now that the URL gets
updated, one can just copy the link from the browser (which is a bit
more intuitive imo, as nothing was really indicating that the clipboard
functionality was there, or that "add to clipboard" was actually what
happened when the name was clicked)
- Gave small visual indication that player names on demon pages can be
  clicked by underlining them.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

* ui: display rank as '-' for unranked players

By using direct stats viewer links, its possible to pull up players that
have no score, and hence no rank. Display rank as '-' for these.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>

---------

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>
Co-authored-by: Patrick <43299462+stadust@users.noreply.github.com>