demonlist-pages: display progress on mobile overview page
Overhaul the mobile list page a bit so that we have the space for the
new progress indicators on there as well. Do this by changing the
thumbnail to take up the full width of the panel, and moving the byline
below it (which is a good change on its own right already imo, because
at leave on my phone the thumbnail's height was significantly less than
the byline's height, and it looked stupid).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>