demonlist-pages: abstract away player-with-statsviewer link rendering
Add a utility struct that renders a given `DatabasePlayer` as a
clickable link to their stats viewer profile.

While we're at it, also convert the player name on the player claim
panel in the list integration tab into one of these, and remove the
button (since we're now using these all over the place, it should be
intuitive enough).

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>