css: define most commonly used colors centrally in variables
Define variables for all most commonly used colors across the website,
unifying a few of them along the way (most motably, record status colors
and info box colors). This should be a decent stepping stone for dark
mode, as all it'll need is moving the variables into a @media
(prefers-color-scheme: [dark|light]), and redefining them appropriately
for dark mode.

Some of the CSS class names will be a bit of a misnomer once dark mode
is a thing (white, blue and red), but I don't want to risk a search and
replace breaking things by not catching all of them, so that's just how
the cookie crumbles.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>