refactor: split login/register into two pages
Turns out that we cannot have two google buttons on the same page, so
the current setup does not allow for a "register with google" flow from
the registration panel. Fix this by simply having login and registration
be different pages. The only UX difference is that swapping between the
two forms isn't as seemless anymore now.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>