ui: css-ify navigation bar frontend
Rely on an almost-pure CSS implementation for the navigation bar. The
old implementation was quite ugly, as it relied on the javascript
cloning a bunch of HTML nodes to create the mobile-only dropdown.
Instead, just do this duplication when generating the HTML (the cloning
of nodes was not cloning event handlers, so depending on the order in
which JS files were loaded, registering event handler to navbar items
was impossible. Or rather, they would only work on the desktop version).
This also eliminates the use of some jQuery, which is always nice. The
cost here are some animations, but really, we don't need those.

While we're at it, remove some unused css definitions. This entire part
of the code pre-dates pointercrate, and stems from when I was messing
around with doing my own Bootstrap-style css framework when I was in
high school, to learn webdev. Good times.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>