fix: do not overwrite listmod-set level_id in -integrate
List Mods can specify level ids for demons, but
GeometryDashConnector::refresh_demon_data will happily overwrite this,
which due to the imperfect heuristic, is not desirable. Fix this by
making sure if a level_id is already set, then we don't try to
re-compute it by searching for the demon name, and instead just use it
straight away.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>