refactor: simplify trp! tokens a bit
Always require a comma after the text_id token, so that the argument
list is a bit easier to understand. This forces us to have a trailing
comma if no arguments are passed, but really, if no arguments are passed
people better be using tr() instead.

While we're at it, change the capture type of $text_id to literal.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>