pages: introduces trp_html! helper macro
This macro is a helper for using the trp! macro in the context of
page-generation, to guarantee we do not accidentally slip in
script-injection opportunities via localization parameters. It
eliminates a ton of usage of maud::PreEscaped and has the invariant that
as long as the input to the macro doesnt use maud::PreEscaped, then
everything will be properly escaped.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>