demonlist-api: simplify Unauthorized handling on GET /players/me
The get_me() function had some unorthodox handling of the authorziation,
since the FromRequest impl for Auth<T> _forwards_ in case of incorrect
authorization. This meant that if authorization is incorrect, it would
move on to GET /players/[player_id], and try to parse "me" as an
integer, and then return 422 UNPROCESSABLE ENTITY. While explicitly
catching the forward to fix this works for now, it's slightly simpler to
just give the /me route lower priority than the [player_id] one, so that
it cannot forward to anything.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>