refactor: librarify geolocation
Do not hardcode the geolocation handler into pointercrate-demonlist-api,
instead allow downstream users of the library to specify their own via
rocket state (by providing an implementation of the GeolocationProvider
trait).

While we're touching this, change the endpoint for geolocation to be
/players/me/geolocate, instead of /players/[player_id]/geolocate. The
latter made no sense, because player_id was forced to be the id of the
claimed player anyway.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>